NetworkX2: Frame Class Complete
The frame class is now fully functional with the ability to produce both position and network videos by overloading th eobject's constructor. All commenting for this class is complete as well. The next step will be to start the video object class which will call the frame class repeatedly to produce a video of our simulation.