Networkx2: Autonomous
By simply creating the FileReader, original Frame, and Video objects we can easily create videos on particle simulation by calling one method. The run method in the Video object now has the option to fully run a specific par/int file and produce the maximum number of frames for a specified experiment. Histogram of the number of  contact/non-contact forces is also plotted.